Organized some code.
