+++ views/mainpanel/meshing_tools/snappyhexmesh.py
@@ -2,8 +2,11 @@ import bpy
 import os
 import json
 from bpy.app.handlers import persistent
-from bpy.types import Operator
-from bpy_extras.io_utils import ExportHelper
+
+# Import operators from their new locations
+from venturial.models.snappyhexmesh.geometry_operators import VNT_OT_create_new_geometry, VNT_OT_delete_geometry
+from venturial.models.snappyhexmesh.file_operators import VNT_OT_export_stl_geometry
+from venturial.models.snappyhexmesh.dictionary_operators import VNT_OT_generate_snappyhex_dict
 
 class snappyhexmesh_menu:
     def layout(self, tools, context):
@@ -31,7 +34,6 @@ class snappyhexmesh_menu:
         cs = context.scene
 
         # STL File section
-        #----------------------------------------------------------------------------------------------
         row = tools.row(align=True)
         row.column(align=True).label(text="Select STL File")
         row.column(align=True).prop(cs, "stl_file", text="") 
@@ -44,20 +46,8 @@ class snappyhexmesh_menu:
         row = tools.row(align=True)
         row.label(text="Name of STL File")
         row.prop(cs, "stl_file_name", text="")
-        #----------------------------------------------------------------------------------------------
-        
-        # Mesh Options section
-        #----------------------------------------------------------------------------------------------
-        box = tools.box()
-        box.label(text="Mesh Options")
-        row = box.row(align=True)
-        row.prop(cs, "castellatedMesh", text="Castellated Mesh")
-        row.prop(cs, "snap", text="Snap")
-        row.prop(cs, "addLayers", text="Add Layers")
-        #----------------------------------------------------------------------------------------------
         
         # Geometry section
-        #----------------------------------------------------------------------------------------------
         row = tools.row(align=True)
         row.column(align=True).label(text="User Defined Geometry")
         row = tools.row(align=True)
@@ -487,214 +477,7 @@ class snappyhexmesh_menu:
         
         # Dictionary actions
         row = box.row(align=True)
-        row.operator("vnt.export_stl_geometry", text="Export STL Geometry")
         row.operator("vnt.generate_snappyhex_dict", text="Generate Dictionary")
-        
-        # Dictionary preview
-        box_dict = tools.box()
-        box_dict.label(text="Generated Dictionary Preview")
-        
-        # Add a preview of the dictionary (partial)
-        if hasattr(cs, 'snappy_dict_preview') and cs.snappy_dict_preview:
-            preview_lines = cs.snappy_dict_preview.split('\n')
-            for i, line in enumerate(preview_lines):
-                if i < 20:  # Show only first 20 lines
-                    box_dict.label(text=line)
-            
-            if len(preview_lines) > 20:
-                box_dict.label(text="... (preview truncated)")
-                
-            # Show total line count
-            box_dict.label(text=f"Total lines: {len(preview_lines)}")
-        else:
-            box_dict.label(text="Click 'Generate Dictionary' to preview the dictionary.")
-
-class VNT_OT_create_new_geometry(Operator):
-    """Create new geometry"""
-    bl_idname = "vnt.create_new_geometry"
-    bl_label = "Create New Geometry"
-
-    geometry_type: bpy.props.EnumProperty(
-        name="Type",
-        description="Select Geometry Type",
-        items=[
-            ("searchableBox", "Box", ""),
-            ("searchableSphere", "Sphere", ""),
-        ],
-        default="searchableBox"
-    )
-    geometry_name: bpy.props.StringProperty( 
-        name="Name",
-        description="Enter the geometry name",
-        default=""
-    )
-    min_x: bpy.props.FloatProperty(name="Min X", default=0.0)
-    min_y: bpy.props.FloatProperty(name="Min Y", default=0.0)
-    min_z: bpy.props.FloatProperty(name="Min Z", default=0.0)
-    max_x: bpy.props.FloatProperty(name="Max X", default=1.0)
-    max_y: bpy.props.FloatProperty(name="Max Y", default=1.0)
-    max_z: bpy.props.FloatProperty(name="Max Z", default=1.0)
-    centre_x: bpy.props.FloatProperty(name="Centre X", default=0.0)
-    centre_y: bpy.props.FloatProperty(name="Centre Y", default=0.0)
-    centre_z: bpy.props.FloatProperty(name="Centre Z", default=0.0)
-    radius: bpy.props.FloatProperty(name="Radius", default=1.0)
-
-    def invoke(self, context, event):
-        return context.window_manager.invoke_props_dialog(self, width=500)
-
-    def draw(self, context):
-        layout = self.layout
-        layout.row().label(text="Name")
-        layout.row().prop(self, "geometry_name", text="")
-        row = layout.row()
-        row.label(text="Type")
-        row.prop(self, "geometry_type", text="")
-        if self.geometry_type == "searchableBox":
-            layout.label(text="Box Settings:")
-            row = layout.row()
-            row.prop(self, "min_x", text="Min X")
-            row.prop(self, "max_x", text="Max X")
-            row = layout.row()
-            row.prop(self, "min_y", text="Min Y")
-            row.prop(self, "max_y", text="Max Y")
-            row = layout.row()
-            row.prop(self, "min_z", text="Min Z")
-            row.prop(self, "max_z", text="Max Z")
-        elif self.geometry_type == "searchableSphere":
-            layout.label(text="Sphere Settings:")
-            row = layout.row()
-            row.prop(self, "centre_x", text="Centre X")
-            row.prop(self, "centre_y", text="Centre Y")
-            row.prop(self, "centre_z", text="Centre Z")
-            layout.prop(self, "radius", text="Radius")
-
-    def execute(self, context):
-        cs = context.scene
-        if self.geometry_type == "searchableBox":
-            width  = self.max_x - self.min_x
-            height = self.max_y - self.min_y
-            depth  = self.max_z - self.min_z
-            center = ((self.min_x + self.max_x) / 2,
-                      (self.min_y + self.max_y) / 2,
-                      (self.min_z + self.max_z) / 2)
-            bpy.ops.mesh.primitive_cube_add(location=center)
-            cube = context.active_object
-            cube.scale = (width / 2, height / 2, depth / 2)
-        elif self.geometry_type == "searchableSphere":
-            loc = (self.centre_x, self.centre_y, self.centre_z)
-            bpy.ops.mesh.primitive_uv_sphere_add(radius=self.radius, location=loc)
-        obj = context.active_object
-        if self.geometry_name:
-            obj.name = self.geometry_name
-        else:
-            obj.name = f"{self.geometry_type}_{len(cs.geometry_items)}"
-        col = bpy.data.collections.get("User Defined Geometry")
-        if not col:
-            col = bpy.data.collections.new("User Defined Geometry")
-            context.scene.collection.children.link(col)
-        if obj.name not in col.objects:
-            col.objects.link(obj)
-        if obj.name in context.scene.collection.objects:
-            context.scene.collection.objects.unlink(obj)
-        new_item = cs.geometry_items.add()
-        new_item.name = obj.name
-        self.report({'INFO'}, f"Created geometry of type {self.geometry_type}")
-        return {'FINISHED'}
-
-class VNT_OT_delete_geometry(Operator):
-    """Delete geometry"""
-    bl_idname = "vnt.delete_geometry"
-    bl_label = "Delete Geometry"
-
-    def execute(self, context):
-        cs = context.scene
-        index = cs.geometry_items_index
-        if index < 0 or index >= len(cs.geometry_items):
-            self.report({'WARNING'}, "Please select geometry to delete")
-            return {'CANCELLED'}
-        geom_name = cs.geometry_items[index].name
-        obj = bpy.data.objects.get(geom_name)
-        if not obj:
-            cs.geometry_items.remove(index)
-            cs.geometry_items_index = min(index, len(cs.geometry_items) - 1)
-            self.report({'WARNING'}, f"Object {geom_name} not found, removed from list")
-            return {'FINISHED'}
-        for col in obj.users_collection:
-            col.objects.unlink(obj)
-        bpy.data.objects.remove(obj)
-        cs.geometry_items.remove(index)
-        cs.geometry_items_index = min(index, len(cs.geometry_items) - 1)
-        self.report({'INFO'}, f"Deleted geometry: {geom_name}")
-        return {'FINISHED'}
-
-class VNT_OT_export_stl_geometry(Operator, ExportHelper):
-    bl_idname = "vnt.export_stl_geometry"
-    bl_label = "Export STL Geometry"
-    
-    filename_ext = ".stl"
-    filter_glob: bpy.props.StringProperty(default="*.stl", options={'HIDDEN'})
-    
-    def execute(self, context):
-        if not context.selected_objects:
-            self.report({'ERROR'}, "No objects selected for export")
-            return {'CANCELLED'}
-        
-        bpy.ops.export_mesh.stl(
-            filepath=self.filepath,
-            check_existing=True,
-            filter_glob="*.stl",
-            use_selection=True,
-            global_scale=1.0,
-            use_scene_unit=False,
-            ascii=False,
-            use_mesh_modifiers=True,
-            batch_mode='OFF'
-        )
-        context.scene.stl_file = self.filepath
-        context.scene.stl_file_name = os.path.basename(self.filepath)
-        
-        self.report({'INFO'}, f"Exported STL to {self.filepath}")
-        return {'FINISHED'}
-
-class VNT_OT_generate_snappyhex_dict(Operator, ExportHelper):
-    """Generate and save a snappyHexMeshDict file"""
-    bl_idname = "vnt.generate_snappyhex_dict"
-    bl_label = "Generate snappyHexMeshDict"
-    
-    filename_ext = ""
-    filter_glob: bpy.props.StringProperty(default="*", options={'HIDDEN'})
-    
-    def execute(self, context):
-        from venturial.models.snappyhexmesh.snappydict_writer import generate_snappy_dict, write_snappy_dict_to_file
-        
-        # Generate dictionary and save to preview
-        dictionary = generate_snappy_dict(context.scene)
-        context.scene.snappy_dict_preview = dictionary
-        
-        if self.filepath:
-            # Make sure directory exists
-            directory = os.path.dirname(self.filepath)
-            if not os.path.exists(directory):
-                try:
-                    os.makedirs(directory)
-                except Exception as e:
-                    self.report({'ERROR'}, f"Failed to create directory: {e}")
-                    return {'CANCELLED'}
-            
-            # Ensure the filepath ends with 'snappyHexMeshDict'
-            target_path = self.filepath
-            if not os.path.basename(target_path) or '.' in os.path.basename(target_path):
-                target_path = os.path.join(target_path, 'snappyHexMeshDict')
-            
-            # Write to file
-            success = write_snappy_dict_to_file(context.scene, target_path)
-            if success:
-                self.report({'INFO'}, f"Dictionary saved to {target_path}")
-            else:
-                self.report({'ERROR'}, f"Failed to write dictionary to {target_path}")
-                return {'CANCELLED'}
-        
-        return {'FINISHED'}
 
 @persistent
 def clean_geometry_items(dummy):
